refactor(subnets): refactored SubnetsList structure
