+++ src/app/subnets/views/Subnets/components/SubnetsTable/constants.ts
