+++ src/app/subnets/views/Subnets/components/SubnetsTable/index.ts
