+++ src/app/subnets/views/Subnets/components/SubnetsTable/types.ts
