+++ src/app/subnets/views/Subnets/components/SubnetsTable/utils.tsx
