+++ src/app/subnets/views/Subnets/views/SubnetsList/_index.scss
