+++ src/app/subnets/views/Subnets/views/SubnetsList/index.ts
