--- src/app/subnets/views/SubnetsList/SubnetsControls/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SubnetsControls";
